requirements.txt
# ── Core orchestration ───────────────────────────────────────────────────────
langchain==0.3.25
langchain-community==0.3.24
langchain-groq==0.3.2

# ── LLM provider ─────────────────────────────────────────────────────────────
groq==0.23.1

# ── Embeddings (local, free) ─────────────────────────────────────────────────
sentence-transformers==3.4.1
torch==2.6.0
huggingface-hub==0.30.2

# ── Vector store ─────────────────────────────────────────────────────────────
faiss-cpu==1.10.0

# ── PDF loading ──────────────────────────────────────────────────────────────
pypdf==5.4.0

# ── Frontend ─────────────────────────────────────────────────────────────────
streamlit==1.45.0

# ── Config & validation ──────────────────────────────────────────────────────
pydantic==2.11.3
pydantic-settings==2.9.1
python-dotenv==1.1.0

# ── Utilities ────────────────────────────────────────────────────────────────
numpy==2.2.4
